LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/disambigaute-forward-expected.txt



FAIL navigation.forward() goes to the nearest forward entry assert_equals: expected 0 but got 1
